pyproject.toml
[tool.poetry]
name = "generate-website-data"
version = "0.1.0"
description = ""
authors = ["Moritz Makowski <moritz.makowski@tum.de>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
pandas = "^1.1.4"
tqdm = "^4.53.0"
faunadb = "^4.0.1"
Cerberus = "^1.3.2"
python-dotenv = "^0.15.0"
firebase-admin = "^5.0.0"
Pillow = "^8.2.0"
rich = "^10.6.0"
scipy = "^1.7.1"
colorcet = "^2.0.6"
geographiclib = "^1.52"
sklearn = "^0.0"
matplotlib = "^3.4.2"
seaborn = "^0.11.1"
bokeh = "^2.3.3"
pyproj = "^3.1.0"
mysql-connector-python = "^8.0.26"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"
pytest = "^6.2.4"
importlib-metadata = "^4.0.1"
typing-extensions = "^3.10.0"
black = "^21.6b0"

[tool.pytest.ini_options]
testpaths = ["tests/"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
